I transformed the global app.js into app.ts a TypeScript file and compiled it into app.js stored in dist file
