+++ package.json
@@ -6,6 +6,5 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "main": "server.js"
-
   }
 }
